IGAPP-901: reverted MapView null check, added useFeatureLocations hook
